chore: Split KameletBinding from routes.ts
