+++ packages/ui/src/models/camel-entities/flows/kamelet.binding.test.ts
@@ -1,26 +1,5 @@
 import { EntityType } from '../base-entity';
-import { CamelRoute, KameletBinding } from './routes';
-
-describe('Camel Route', () => {
-  let camelEntity: CamelRoute;
-
-  beforeEach(() => {
-    camelEntity = new CamelRoute();
-  });
-
-  it('should have an uuid', () => {
-    expect(camelEntity.id).toBeDefined();
-    expect(typeof camelEntity.id).toBe('string');
-  });
-
-  it('should have a type', () => {
-    expect(camelEntity.type).toEqual(EntityType.Route);
-  });
-
-  it('should return the steps', () => {
-    expect(camelEntity.getSteps()).toEqual([]);
-  });
-});
+import { KameletBinding } from './kamelet-binding';
 
 describe('Kamelet Binding', () => {
   let kameletBinding: KameletBinding;
