+++ packages/ui/src/models/camel-entities/flows/route.test.ts
@@ -1,0 +1,23 @@
+import { EntityType } from '../base-entity';
+import { CamelRoute } from './route';
+
+describe('Camel Route', () => {
+  let camelEntity: CamelRoute;
+
+  beforeEach(() => {
+    camelEntity = new CamelRoute();
+  });
+
+  it('should have an uuid', () => {
+    expect(camelEntity.id).toBeDefined();
+    expect(typeof camelEntity.id).toBe('string');
+  });
+
+  it('should have a type', () => {
+    expect(camelEntity.type).toEqual(EntityType.Route);
+  });
+
+  it('should return the steps', () => {
+    expect(camelEntity.getSteps()).toEqual([]);
+  });
+});
