+++ packages/ui/src/models/camel-entities/index.ts
@@ -1,4 +1,4 @@
 export * from './base-entity';
 
 /** Routes definitions */
-export * from './flows/routes';
+export * from './flows';
